api/requirements.txt
fastapi
uvicorn[standard]
mongoengine
pymongo
pydantic[email]
python-jose[cryptography]
bcrypt
python-dotenv
